classes fix
